+++ dbs-21/src/Dash.css
@@ -1,0 +1,46 @@
+.app {
+    padding: 20px;
+    background-color: #f9f9f9; /* or any color you prefer */
+  }
+  
+  .dashboard {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: 20px;
+  }
+  
+  .card {
+    border: 1px solid #ddd;
+    border-radius: 10px;
+    width: 250px;
+    margin: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease-in-out;
+    background-color: #fff; /* card background color */
+  }
+  
+  .card:hover {
+    transform: translateY(-5px);
+  }
+  
+  .card-image-container {
+    text-align: center;
+    padding: 10px;
+    background-color: #f5f5f5; /* placeholder color for the image container */
+  }
+  
+  .card-image {
+    max-width: 100%;
+    height: auto;
+    border-radius: 8px; /* Optional: if you want rounded corners on the image */
+  }
+  
+  .card-description {
+    padding: 15px;
+  }
+  
+  h1{
+    text-align: center;
+  
+  }
